Worked on converting unit tests to the componentRunner scaffolding
