+++ src/test/test.ts
@@ -12,7 +12,7 @@ import componentMount from "./componentMount.js";
 import nodeState from "./nodeState/index.js";
 import replaceByKey from "./replaceByKey/index.js";
 import clearsOldProps from "./clearsOldProps/index.js";
-import shouldUpdate from "./shouldUpdate/index.js";
+import shouldUpdate from "./shouldUpdate.js";
 import renderPrimitives from "./renderPrimitives/index.js";
 import assertIsComponent from "./assertIsComponent/index.js";
 import rendersArraysInChildren from "./rendersArraysInChildren/index.js";
@@ -25,17 +25,17 @@ import mountRunsOnceWhenChildRendersFragment from "./mountRunsOnceWhenChildRende
 import fragmentUnmountRunsOnce from "./fragmentUnmountRunsOnce/index.js";
 import replacingFragmentWithNodeWontUnmount from "./replacingFragmentWithNodeWontUnmount/index.js";
 import childWithFragmentUnmounts from "./childWithFragmentUnmounts/index.js";
-import rerenderMayUnmountParents from "./rerenderMayUnmountParents/index.js";
-import rerenderMayChangeRootNode from "./rerenderMayChangeRootNode/index.js";
+import rerenderMayUnmountParents from "./rerenderMayUnmountParents.js";
+import rerenderMayChangeRootNode from "./rerenderMayChangeRootNode.js";
 import dangerouslySetInnerHTML from "./dangerouslySetInnerHTML/index.js";
 import css from "./css/index.js";
 import inheritedCustomElement from "./inheritedCustomElement/index.js";
 import fragmentOverwriteDoesNotUnmount from "./fragmentOverwriteDoesNotUnmount/index.js";
-import ssrSimple from "./ssr-simple/index.js";
+import ssrSimple from "./ssr-simple.js";
 import componentKeepsStateWhenReordered from "./componentKeepsStateWhenReordered/index.js";
 import elementKeepsStateWhenReordered from "./elementKeepsStateWhenReordered/index.js";
 import unmountsParentWhenNodeIsNull from "./unmountsParentWhenNodeIsNull.js";
-import rerenderMayChangeRootNodeOnParents from "./rerenderMayChangeRootNodeOnParents/index.js";
+import rerenderMayChangeRootNodeOnParents from "./rerenderMayChangeRootNodeOnParents.js";
 import keyedFragmentsPreserveChildStates from "./keyedFragmentsPreserveChildStates/index.js";
 import rootElementChangeDoesNotUnmount from "./rootElementChangeDoesNotUnmount.js";
 import fragmentMountTiming from "./fragmentMountEvent.js";
